pyproject.toml
[project]
name = "nonebot-plugin-aiochatllm"
version = "0.1.0"
description = "All-in-One Large Model Chat Plugin"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Raven95676", email = "raven95676@gmail.com"},
]
license = {text = "MIT"}
dependencies = [
    "nonebot2[fastapi,httpx]>=2.4.1", 
    "nonebot-plugin-alconna>=0.55.1", 
    "nonebot-plugin-localstore>=0.7.4", 
    "nonebot-plugin-uninfo>=0.6.10", 
    "httpx>=0.28.1", 
    "chromadb>=0.6.3"
]

[tool.nonebot]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" },
    { name = "Telegram", module_name = "nonebot.adapters.telegram" }
]
plugins = ["nonebot_plugin_alconna","nonebot_plugin_localstore","nonebot_plugin_uninfo"]
plugin_dirs = ["src"]

[tool.pdm]

[tool.pdm.build]
includes = ["src/nonebot_plugin_aiochatllm"]

[tool.ruff]
line-length = 120
target-version = "py310"
select = ["E", "F", "I", "W", "N", "B", "C4", "UP", "ANN", "RUF"]
ignore = ["ANN101","ANN401","RUF001"]
unfixable = ["F401"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
[dependency-groups]
dev = [
    "nonebot-adapter-onebot>=2.4.6",
    "nonebot-adapter-telegram>=0.1.0b20",
    "nonebot-adapter-console>=0.6.0",
    "nonebot-adapter-mail>=1.0.0a7",
    "mypy>=1.15.0",
    "pytest>=8.3.4",
    "ruff>=0.3.0",
    "nonebug>=0.4.3",
]
